crawler/requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify ~= 1.7.0
psycopg2-binary>=2.9.3
beautifulsoup4==4.10.0
requests==2.31.0
apify_client
selenium
webdriver_manager
